Expand physics/__init__.py to evolve and utils
