+++ pooltool/physics/resolve/resolver.py
@@ -1,3 +1,4 @@
+"""Resolve collisions and transitions"""
 from __future__ import annotations
 
 from pathlib import Path
